+++ src/db.json
@@ -13,7 +13,25 @@
       "postId": 1
     }
   ],
-  "profile": {
-    "name": "typicode"
-  }
+  "profile": [
+    {
+      "name": "typicode"
+    }
+  ],
+  "boards": [
+    {
+      "postid": 65465456465465,
+      "title": "The Cloud Waste Problem",
+      "contents": "With the public cloud adoption increasing year over year, companies will spend nearly $121.6 billion in IaaS (Infrastructure as a Service) in 2022 from AWS, Azure, Google, and other cloud service providersh111\n\n",
+      "id": 1
+    }
+  ],
+  "boardsreply": [
+    {
+      "sameId": 2,
+      "comment": "test2",
+      "like": 0,
+      "id": 4
+    }
+  ]
 }
